+++ tests/test_create_model.py
@@ -61,6 +61,17 @@ def test_custom_config():
         model(foo=654)
 
 
+def test_custom_config_extras():
+    class Config(BaseModel.Config):
+        ignore_extra = False
+        allow_extra = False
+
+    model = create_model('FooModel', foo=(int, ...), __config__=Config)
+    assert model(foo=654)
+    with pytest.raises(ValidationError):
+        model(bar=654)
+
+
 def test_inheritance_validators():
     class BarModel(BaseModel):
         @validator('a', check_fields=False)
@@ -71,6 +82,23 @@ def test_inheritance_validators():
 
     model = create_model('FooModel', a='cake', __base__=BarModel)
     assert model().a == 'cake'
+    assert model(a='this is foobar good').a == 'this is foobar good'
+    with pytest.raises(ValidationError):
+        model(a='something else')
+
+
+def test_inheritance_validators_always():
+    class BarModel(BaseModel):
+        @validator('a', check_fields=False, always=True)
+        def check_a(cls, v):
+            if 'foobar' not in v:
+                raise ValueError('"foobar" not found in a')
+            return v
+
+    model = create_model('FooModel', a='cake', __base__=BarModel)
+    with pytest.raises(ValidationError):
+        model()
+    assert model(a='this is foobar good').a == 'this is foobar good'
     with pytest.raises(ValidationError):
         model(a='something else')